Move all files default files into subfolders
