+++ src/t8_schemes/t8_default/t8_default_common/t8_default_common_cxx.cxx
@@ -21,7 +21,7 @@
 */
 
 #include <sc_functions.h>
-#include "t8_default_common_cxx.hxx"
+#include <t8_schemes/t8_default/t8_default_common/t8_default_common_cxx.hxx>
 
 /* We want to export the whole implementation to be callable from "C" */
 T8_EXTERN_C_BEGIN ();
