+++ src/t8_schemes/t8_default/t8_default_line/t8_default_line_cxx.cxx
@@ -20,11 +20,11 @@
   51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 */
 
-#include "t8_default_common_cxx.hxx"
-#include "t8_default_vertex_cxx.hxx"
-#include "t8_dvertex_bits.h"
-#include "t8_dline_bits.h"
-#include "t8_dline.h"
+#include <t8_schemes/t8_default/t8_default_common/t8_default_common_cxx.hxx>
+#include <t8_schemes/t8_default/t8_default_vertex/t8_default_vertex_cxx.hxx>
+#include <t8_schemes/t8_default/t8_default_vertex/t8_dvertex_bits.h>
+#include <t8_schemes/t8_default/t8_default_line/t8_dline_bits.h>
+#include <t8_schemes/t8_default/t8_default_line/t8_dline.h>
 
 typedef t8_dline_t  t8_default_line_t;
