+++ src/t8_schemes/t8_default/t8_default_line/t8_dline_bits.h
@@ -27,8 +27,8 @@
 #define T8_DLINE_BITS_H
 
 #include <t8_element.h>
-#include "t8_dline.h"
-#include "t8_dvertex.h"
+#include <t8_schemes/t8_default/t8_default_line/t8_dline.h>
+#include <t8_schemes/t8_default/t8_default_vertex/t8_dvertex.h>
 
 T8_EXTERN_C_BEGIN ();
