+++ src/t8_schemes/t8_default/t8_default_prism/t8_dprism.h
@@ -28,8 +28,8 @@
  */
 
 #include <t8.h>
-#include "t8_dline.h"
-#include "t8_dtri.h"
+#include <t8_schemes/t8_default/t8_default_line/t8_dline.h>
+#include <t8_schemes/t8_default/t8_default_tri/t8_dtri.h>
 
 /** The number of children that a prism is refined into. */
 #define T8_DPRISM_CHILDREN 8
