+++ src/t8_schemes/t8_default/t8_default_quad/t8_default_quad_cxx.cxx
@@ -21,9 +21,9 @@
 */
 
 #include <p4est_bits.h>
-#include "t8_dline_bits.h"
-#include "t8_default_common_cxx.hxx"
-#include "t8_default_quad_cxx.hxx"
+#include <t8_schemes/t8_default/t8_default_line/t8_dline_bits.h>
+#include <t8_schemes/t8_default/t8_default_common/t8_default_common_cxx.hxx>
+#include <t8_schemes/t8_default/t8_default_quad/t8_default_quad_cxx.hxx>
 
 /* We want to export the whole implementation to be callable from "C" */
 T8_EXTERN_C_BEGIN ();
