+++ src/t8_schemes/t8_default/t8_default_quad/t8_default_quad_cxx.hxx
@@ -33,8 +33,8 @@
 
 #include <p4est.h>
 #include <t8_element_cxx.hxx>
-#include "t8_default_line_cxx.hxx"
-#include "t8_default_common_cxx.hxx"
+#include <t8_schemes/t8_default/t8_default_line/t8_default_line_cxx.hxx>
+#include <t8_schemes/t8_default/t8_default_common/t8_default_common_cxx.hxx>
 
 /** The structure holding a quadrilateral element in the default scheme.
  * We make this definition public for interoperability of element classes.
