+++ src/t8_schemes/t8_default/t8_default_tet/t8_dtet_bits.c
@@ -20,8 +20,8 @@
   51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 */
 
-#include "t8_dtri_to_dtet.h"
-#include "t8_dtri_bits.c"
+#include <t8_schemes/t8_default/t8_default_tet/t8_dtri_to_dtet.h>
+#include <t8_schemes/t8_default/t8_default_tri/t8_dtri_bits.c>
 
 int
 t8_dtet_is_familypv (const t8_dtri_t * f[])
