+++ src/t8_schemes/t8_default/t8_default_tri/t8_default_tri_cxx.hxx
@@ -20,31 +20,26 @@
   51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 */
 
-/** \file t8_default_hex.h
+/** \file t8_default_tri.h
+ * The default implementation for triangles.
  */
 
-#ifndef T8_DEFAULT_HEX_HXX
-#define T8_DEFAULT_HEX_HXX
+#ifndef T8_DEFAULT_TRI_CXX_H
+#define T8_DEFAULT_TRI_CXX_H
 
-#include <p8est.h>
 #include <t8_element_cxx.hxx>
-#include "t8_default_quad_cxx.hxx"
+#include <t8_schemes/t8_default/t8_default_line/t8_default_line_cxx.hxx>
+#include <t8_schemes/t8_default/t8_default_common/t8_default_common_cxx.hxx>
 
-/** The structure holding a hexahedral element in the default scheme.
- * We make this definition public for interoperability of element classes.
- * We might want to put this into a private, scheme-specific header file.
- */
-typedef p8est_quadrant_t t8_phex_t;
-
-struct t8_default_scheme_hex_c:public t8_default_scheme_common_c
+struct t8_default_scheme_tri_c:public t8_default_scheme_common_c
 {
 public:
   /** The virtual table for a particular implementation of an element class. */
 
   /** Constructor. */
-  t8_default_scheme_hex_c ();
+  t8_default_scheme_tri_c ();
 
-  ~t8_default_scheme_hex_c ();
+  ~t8_default_scheme_tri_c ();
 
   /** Allocate memory for a given number of elements.
    * In debugging mode, ensure that all elements are valid \ref t8_element_is_valid.
@@ -59,7 +54,11 @@ public:
   virtual int         t8_element_maxlevel (void);
 
 /** Return the type of each child in the ordering of the implementation. */
-  virtual t8_eclass_t t8_element_child_eclass (int childid);
+  virtual t8_eclass_t t8_element_child_eclass (int childid)
+  {
+    SC_ABORT ("This function is not implemented yet.\n");
+    return T8_ECLASS_ZERO;      /* suppresses compiler warning */
+  }
 
 /** Return the refinement level of an element. */
   virtual int         t8_element_level (const t8_element_t * elem);
@@ -100,20 +99,16 @@ public:
   virtual int         t8_element_num_face_children (const t8_element_t *
                                                     elem, int face);
 
+  /** Return the corner number of an element's face corner. */
   virtual int         t8_element_get_face_corner (const t8_element_t *
                                                   element, int face,
                                                   int corner);
 
-  /** Return the face numbers of the faces sharing an element's corner. */
   virtual int         t8_element_get_corner_face (const t8_element_t *
                                                   element, int corner,
-                                                  int face)
-  {
-    SC_ABORT ("Not implemented.\n");
-    return 0;                   /* prevents compiler warning */
-  }
+                                                  int face);
 
-/** Construct the child element of a given number. */
+/** Construct the child element of a given number (in tetrahedral Morton order). */
   virtual void        t8_element_child (const t8_element_t * elem,
                                         int childid, t8_element_t * child);
 
@@ -136,7 +131,7 @@ public:
                                       const t8_element_t * elem2,
                                       t8_element_t * nca);
 
-  /** Compute the elmement class of the face of an element. */
+  /** Compute the element shape of the face of an element. */
   virtual t8_element_shape_t t8_element_face_shape (const t8_element_t * elem,
                                                     int face);
 
@@ -150,7 +145,8 @@ public:
                                                    int num_children,
                                                    int *child_indices);
 
-  /** Given a face of an element and a child number of a child of that face, return the face number
+  /** Given a face of an element and a child number (in Morton order)
+   *  of a child of that face, return the face number
    * of the child of the element that matches the child face. */
   virtual int         t8_element_face_child_face (const t8_element_t * elem,
                                                   int face, int face_child);
@@ -165,16 +161,12 @@ public:
   virtual int         t8_element_tree_face (const t8_element_t * elem,
                                             int face);
 
-  /** Transform the coordinates of a hexahedron considered as boundary element
+  /** Transform the coordinates of a triangle considered as boundary element
    *  in a tree-tree connection. */
   virtual void        t8_element_transform_face (const t8_element_t * elem1,
                                                  t8_element_t * elem2,
-                                                 int orientation,
-                                                 int sign,
-                                                 int is_smaller_face)
-  {
-    SC_ABORT ("This function is not implemented yet.\n");
-  }
+                                                 int orientation, int sign,
+                                                 int is_smaller_face);
 
   /** Given a boundary face inside a root tree's face construct
    *  the element inside the root tree that has the given face as a
@@ -202,8 +194,8 @@ public:
   virtual void        t8_element_boundary_face (const t8_element_t * elem,
                                                 int face,
                                                 t8_element_t * boundary,
-                                                const t8_eclass_scheme_c *
-                                                boundary_scheme);
+                                                const t8_eclass_scheme_c
+                                                * boundary_scheme);
 
 /** Construct all codimension-one boundary elements of a given element. */
   virtual void        t8_element_boundary (const t8_element_t * elem,
@@ -266,6 +258,17 @@ public:
   virtual void        t8_element_vertex_coords (const t8_element_t * t,
                                                 int vertex, int coords[]);
 
+  /** The triangle schemes uses the general function to return the type of
+   * a triangle.
+   *  \param [in] elem An valid element
+   *  \param [in] indata Is ignored. Can be NULL.
+   *  \param [out] outdata Pointer to an int8_t. The type of \a elem will be stored here.
+   *  On output the type of the triangle will be stored in \a outdata
+   */
+  virtual void        t8_element_general_function (const t8_element_t * elem,
+                                                   const void *indata,
+                                                   void *outdata);
+                                                   
   /** Compute the coordinates of a given element vertex inside a reference tree
    *  that is embedded into [0,1]^d (d = dimension).
    *   \param [in] t      The element to be considered.
@@ -283,4 +286,4 @@ public:
 #endif
 };
 
-#endif /* !T8_DEFAULT_HEX_HXX */
+#endif /* !T8_DEFAULT_TET_H */
