+++ src/t8_schemes/t8_default/t8_default_tri/t8_dtri_bits.c
@@ -21,11 +21,11 @@
 */
 
 #ifndef T8_DTRI_TO_DTET
-#include "t8_dtri_bits.h"
-#include "t8_dtri_connectivity.h"
+#include <t8_schemes/t8_default/t8_default_tri/t8_dtri_bits.h>
+#include <t8_schemes/t8_default/t8_default_tri/t8_dtri_connectivity.h>
 #else
-#include "t8_dtet_bits.h"
-#include "t8_dtet_connectivity.h"
+#include <t8_schemes/t8_default/t8_default_tet/t8_dtet_bits.h>
+#include <t8_schemes/t8_default/t8_default_tet/t8_dtet_connectivity.h>
 #endif
 
 typedef int8_t      t8_dtri_cube_id_t;
