+++ src/t8_schemes/t8_default/t8_default_vertex/t8_dvertex_bits.c
